Implements repository pattern to allow multiple storage for metadata
